+++ common/contract/chaintype.go
@@ -4,8 +4,8 @@ package contract
 type ChainType interface {
 	// GetChainTablePrefix return the value of current chain table prefix in the database
 	GetTablePrefix() string
-	// GetChainForgingDelayTime return the value of chain forging delay in second
-	GetChainSmithingDelayTime() uint32
+	// GetChainSmithingDelayTime return the value of chain smithing delay in second
+	GetChainSmithingDelayTime() int64
 	// GetName return the name of the chain : used in parsing chaintype across node
 	GetName() string
 	// GetGenesisBlockID return the block ID of genesis block in the chain
